node: remove all references to nodejs
nodejs is too flaky. Remove it and assume the source files are
preprocessed prior to compression.

In practice, we lose about 8 kB. However, this is because currently our
CSS is unminified. This can be recouped by minifying it in the UI build.

As a tradeoff, the installation process becomes quicker and we don't
have to worry about the deluge of npm error messages.

Signed-off-by: Sean Cross <sean@xobs.io>